move all modules to a single folder
